refactor: suggestController 분리
