+++ server/src/test/java/com/codueon/boostUp/domain/suggest/toss/TossPaymentTest.java
@@ -1,14 +1,10 @@
-package com.codueon.boostUp.domain.suggest.contoller;
+package com.codueon.boostUp.domain.suggest.toss;
 
 import com.codueon.boostUp.domain.suggest.response.Message;
-import com.codueon.boostUp.domain.suggest.toss.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.test.web.servlet.ResultActions;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import static com.codueon.boostUp.domain.suggest.utils.SuggestConstants.*;
 import static com.codueon.boostUp.domain.utils.ApiDocumentUtils.getRequestPreProcessor;
@@ -25,7 +21,7 @@ import static org.springframework.restdocs.request.RequestDocumentation.pathPara
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-public class TossPaymentTest extends SuggestControllerTest{
+public class TossPaymentTest extends TossPayControllerTest {
 
     @Test
     @DisplayName("GET 신청 프로세스 4-2 Toss 결제 URL 요청")
@@ -45,7 +41,7 @@ public class TossPaymentTest extends SuggestControllerTest{
                 .message(TOSS_PAY_URI_MSG)
                 .build();
 
-        given(suggestService.getTossPayUrl(Mockito.anyLong(), Mockito.anyInt()))
+        given(tossPayService.getTossPayUrl(Mockito.anyLong(), Mockito.anyInt()))
                 .willReturn(message);
 
         ResultActions actions =
@@ -108,7 +104,7 @@ public class TossPaymentTest extends SuggestControllerTest{
                 .message(INFO_URI_MSG)
                 .build();
 
-        given(suggestService.getSuccessTossPaymentInfo(Mockito.anyLong()))
+        given(tossPayService.getSuccessTossPaymentInfo(Mockito.anyLong()))
                 .willReturn(message);
 
         ResultActions actions =
@@ -143,70 +139,4 @@ public class TossPaymentTest extends SuggestControllerTest{
                         )
                 ));
     }
-
-    @Test
-    @DisplayName("GET 신청 프로세스 9-2 Toss 환불")
-    void refundPaymentKakaoOrToss() throws Exception {
-        Integer count = 2;
-        Integer amount = 5000;
-
-        RequestForTossPayCancelInfo body = RequestForTossPayCancelInfo.builder()
-                .cancelReason("고객이 취소를 요청함")
-                .cancelAmount(count * amount)
-                .build();
-
-        List<Cancels> cancels = new ArrayList<>();
-        Checkout checkout = Checkout.builder().build();
-        TossCard card = TossCard.builder().build();
-        Receipt receipt = Receipt.builder().build();
-        MobilePhone mobilePhone = MobilePhone.builder().build();
-        Transfer transfer = Transfer.builder().build();
-
-        TossPayCancelInfo cancelInfo = TossPayCancelInfo.builder()
-                .cancels(cancels)
-                .totalAmount(count * amount)
-                .paymentKey("paymentKey")
-                .lastTransactionKey("lastTransactionKey")
-                .method("method")
-                .orderId("orderId")
-                .orderName("orderName")
-                .checkout(checkout)
-                .mId("mId")
-                .requestedAt("requestedAt")
-                .approvedAt("approvedAt")
-                .card(card)
-                .receipt(receipt)
-                .mobilePhone(mobilePhone)
-                .transfer(transfer)
-                .build();
-
-        Message message = Message.builder()
-                .data(cancelInfo)
-                .message(CANCELED_PAY_MESSAGE)
-                .build();
-
-        given(suggestService.refundPaymentKakaoOrToss(Mockito.anyLong(), Mockito.anyLong()))
-                .willReturn(message);
-
-        ResultActions actions =
-                mockMvc.perform(
-                        get("/suggest/{suggest-id}/refund", suggest.getId())
-                                .header(AUTHORIZATION, BEARER + accessToken)
-                                .header(REFRESH_TOKEN, refreshToken)
-                );
-
-        actions
-                .andExpect(status().isOk())
-                .andDo(document("신청9.2-토스결제환불",
-                        getRequestPreProcessor(),
-                        pathParameters(
-                                parameterWithName("suggest-id").description("신청 식별자")
-                        ),
-                        requestHeaders(
-                                headerWithName(AUTHORIZATION).description("엑세스 토큰"),
-                                headerWithName(REFRESH_TOKEN).description("리프레시 토큰")
-                        )
-                ));
-    }
-
 }
